requirements.txt
# Core Django
Django>=4.2.0,<5.0.0
djangorestframework>=3.14.0,<4.0.0
django-cors-headers>=4.0.0,<5.0.0
djangorestframework-simplejwt>=5.2.0,<6.0.0

# Database
psycopg2-binary>=2.9.0,<3.0.0
dj-database-url>=2.0.0,<3.0.0

# Production server
gunicorn>=20.0.0,<22.0.0
whitenoise>=6.0.0,<7.0.0

# AI/ML dependencies  
openai>=1.50.0,<2.0.0
aiohttp>=3.8.0,<4.0.0

# Image processing
pillow>=10.0.0,<12.0.0

# Barcode detection
opencv-python>=4.8.0,<5.0.0
pyzbar>=0.1.9,<1.0.0

# API integration
requests>=2.28.0,<3.0.0

# Configuration
python-decouple>=3.6,<4.0
python-dotenv>=1.0.0,<2.0.0

# Utilities
pytz>=2023.3
asgiref>=3.6.0,<4.0.0
sqlparse>=0.4.0,<1.0.0
psutil==7.0.0